requirements.txt
# DocBot Dependencies

# Environment variables
python-dotenv>=1.0.0

# LLM - Universal interface for multiple providers
litellm>=1.30.0

# Vector database
chromadb>=0.4.22

# Telegram bot
python-telegram-bot>=21.0

# Discord bot
discord.py>=2.3.2

# HTTP requests
httpx>=0.27.0

# PDF support (optional)
pypdf>=3.17.0
